chore(server): gendo IoC #1 - prep/cleanup before getting started
