+++ packages/frontend-2/lib/gendo/graphql/queriesAndMutations.ts
@@ -1,5 +1,7 @@
 import { graphql } from '~~/lib/common/generated/gql'
 
+// TODO: Clean up these operations and make them component fragment based. Also some of the props requested don't seem to even be used
+
 export const requestGendoAIRender = graphql(`
   mutation requestGendoAIRender($input: GendoAIRenderInput!) {
     versionMutations {
